backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-nested-routers==0.93.4
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
